add some more adapter patterns
