+++ src/org/smof/collection/CollectionsPoolImpl.java
@@ -1,0 +1,116 @@
+/*
+ *******************************************************************************
+ * Copyright (C) 2017 Joao
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ *****************************************************************************
+ */
+package org.smof.collection;
+
+import java.util.Iterator;
+import java.util.Map;
+
+import org.smof.element.Element;
+import org.smof.exception.NoSuchCollection;
+import org.smof.exception.SmofException;
+
+import com.google.common.collect.Maps;
+import com.mongodb.client.gridfs.GridFSBucket;
+
+@SuppressWarnings("javadoc")
+public class CollectionsPoolImpl implements CollectionsPool {
+	
+	private static void handleError(Throwable cause) {
+		throw new SmofException(cause);
+	}
+	
+	private final Map<Class<? extends Element>, SmofCollection<? extends Element>> collections;
+	private final Map<String, GridFSBucket> fsBuckets;
+	
+	public CollectionsPoolImpl() {
+		collections = Maps.newLinkedHashMap();
+		fsBuckets = Maps.newLinkedHashMap();
+	}
+	
+	@Override
+    public void put(String bucketName, GridFSBucket bucket) {
+		fsBuckets.put(bucketName, bucket);
+	}
+	
+	@Override
+    public <T extends Element> void put(Class<T> elClass, SmofCollection<T> collection) {
+		collections.put(elClass, collection);
+	}
+	
+	@Override
+    @SuppressWarnings("unchecked")
+	public <T extends Element> SmofCollection<T> getCollection(Class<T> elClass) {
+		final Class<? extends Element> validSuperType = getValidSuperType(elClass);
+		if(validSuperType == null) {
+			handleError(new NoSuchCollection(elClass));
+			return null;
+		}
+		return (SmofCollection<T>) collections.get(validSuperType);
+	}
+	
+	@Override
+    public GridFSBucket getBucket(String bucketName) {
+		return fsBuckets.get(bucketName);
+	}
+	
+	@Override
+    public Iterable<String> getAllBuckets() {
+		return fsBuckets.keySet();
+	}
+	
+	@Override
+    public void removeBucket(String bucketName) {
+		fsBuckets.remove(bucketName);
+	}
+	
+	@Override
+    public void clearBuckets() {
+		fsBuckets.clear();
+	}
+
+	@Override
+	public Iterator<SmofCollection<?>> iterator() {
+		return collections.values().iterator();
+	}
+
+	private Class<? extends Element> getValidSuperType(Class<? extends Element> elClass) {
+		for(Class<? extends Element> type : collections.keySet()) {
+			if(type.isAssignableFrom(elClass)) {
+				return type;
+			}
+		}
+		return null;
+	}
+
+	@Override
+    public void remove(SmofCollection<?> collection) {
+		collections.remove(collection.getType());
+	}
+	
+	@Override
+    public void clearCollections() {
+		collections.clear();
+	}
+
+}
