+++ src/org/smof/collection/SmofOpOptions.java
@@ -1,153 +1,27 @@
-/*******************************************************************************
- * Copyright (C) 2017 Joao
- * 
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- * 
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- * 
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- ******************************************************************************/
 package org.smof.collection;
-
-import java.util.concurrent.TimeUnit;
-
-import org.apache.commons.lang3.tuple.Pair;
-import org.bson.conversions.Bson;
 
 import com.mongodb.client.model.FindOneAndReplaceOptions;
 import com.mongodb.client.model.FindOneAndUpdateOptions;
 import com.mongodb.client.model.ReturnDocument;
 import com.mongodb.client.model.UpdateOptions;
 
-@SuppressWarnings("javadoc")
-public class SmofOpOptions {
-	
-	public static SmofOpOptions create() {
-		return new SmofOpOptions();
-	}
-	
-	private boolean upsert;
-	private boolean validateDocuments;
-	private Pair<Long, TimeUnit> maxTime;
-	private Bson projection;
-	private Bson sort;
-	private ReturnDocument ret;
-	private boolean bypassCache;
-	
-	private SmofOpOptions() {
-		upsert = false;
-		validateDocuments = true;
-		ret = ReturnDocument.AFTER;
-		bypassCache = false;
-	}
-	
-	public FindOneAndReplaceOptions toFindOneAndReplace() {
-		final FindOneAndReplaceOptions options = new FindOneAndReplaceOptions();
-		options.bypassDocumentValidation(!validateDocuments);
-		setMaxTime(options);
-		setProjection(options);
-		options.returnDocument(ret);
-		setSort(options);
-		options.upsert(upsert);
-		return options;
-	}
+public interface SmofOpOptions {
+    FindOneAndReplaceOptions toFindOneAndReplace();
 
-	private void setSort(final FindOneAndReplaceOptions options) {
-		if(sort != null) {
-			options.sort(sort);
-		}
-	}
+    UpdateOptions toUpdateOptions();
 
-	private void setProjection(final FindOneAndReplaceOptions options) {
-		if(projection != null) {
-			options.projection(projection);
-		}
-	}
+    FindOneAndUpdateOptions toFindOneAndUpdateOptions();
 
-	private void setMaxTime(final FindOneAndReplaceOptions options) {
-		if(maxTime != null) {
-			options.maxTime(maxTime.getLeft(), maxTime.getRight());
-		}
-	}
-	
-	public UpdateOptions toUpdateOptions() {
-		final UpdateOptions options = new UpdateOptions();
-		options.bypassDocumentValidation(!validateDocuments);
-		options.upsert(upsert);
-		return options;
-	}
-	
-	public FindOneAndUpdateOptions toFindOneAndUpdateOptions() {
-		final FindOneAndUpdateOptions options = new FindOneAndUpdateOptions();
-		options.bypassDocumentValidation(!validateDocuments);
-		setMaxTime(options);
-		setProjection(options);
-		options.returnDocument(ret);
-		setSort(options);
-		options.upsert(upsert);
-		return options;
-	}
+    void upsert(boolean upsert);
 
-	private void setSort(final FindOneAndUpdateOptions options) {
-		if(sort != null) {
-			options.sort(sort);
-		}
-	}
+    void validateDocuments(boolean validateDocuments);
 
-	private void setProjection(final FindOneAndUpdateOptions options) {
-		if(projection != null) {
-			options.projection(projection);
-		}
-	}
+    void setReturnDocument(ReturnDocument doc);
 
-	private void setMaxTime(final FindOneAndUpdateOptions options) {
-		if(maxTime != null) {
-			options.maxTime(maxTime.getLeft(), maxTime.getRight());
-		}
-	}
-	
-	public void upsert(boolean upsert) {
-		this.upsert = upsert;
-	}
-	
-	public void validateDocuments(boolean validateDocuments) {
-		this.validateDocuments = validateDocuments;
-	}
-	
-	public void setReturnDocument(ReturnDocument doc) {
-		this.ret = doc;
-	}
-	
-	public void bypassCache(boolean bypassCache) {
-		this.bypassCache = bypassCache;
-	}
-	
-	public boolean isBypassCache() {
-		return bypassCache;
-	}
+    void bypassCache(boolean bypassCache);
 
-	@Override
-	public String toString() {
-		return "SmofOpOptions [upsert=" + upsert 
-				+ ", validateDocuments=" + validateDocuments 
-				+ ", maxTime=" + maxTime
-				+ ", projection=" + projection 
-				+ ", sort=" + sort 
-				+ ", ret=" + ret 
-				+ ", bypassCache=" + bypassCache
-				+ "]";
-	}
-	
+    boolean isBypassCache();
+
+    @Override
+    String toString();
 }
